Create shared PaymentConfigurationModule to deduplicate Dagger providers
The same 3 providers (PaymentConfiguration, publishableKey, stripeAccountId) were
copy-pasted across 10+ Dagger modules. Extract them into a single shared module in
payments-core to reduce duplication and ensure consistent behavior.

Committed-By-Agent: claude
