+++ payment-method-messaging/src/main/java/com/stripe/android/paymentmethodmessaging/element/PaymentMethodMessagingModule.kt
@@ -3,9 +3,7 @@ package com.stripe.android.paymentmethodmessaging.element
 import android.app.Application
 import android.content.Context
 import com.stripe.android.BuildConfig
-import com.stripe.android.PaymentConfiguration
 import com.stripe.android.core.injection.ENABLE_LOGGING
-import com.stripe.android.core.injection.PUBLISHABLE_KEY
 import com.stripe.android.core.injection.ViewModelScope
 import com.stripe.android.core.networking.AnalyticsRequestFactory
 import com.stripe.android.core.utils.DefaultDurationProvider
@@ -16,6 +14,7 @@ import com.stripe.android.paymentmethodmessaging.element.analytics.PaymentMethod
 import com.stripe.android.payments.core.analytics.ErrorReporter
 import com.stripe.android.payments.core.analytics.RealErrorReporter
 import com.stripe.android.payments.core.injection.PRODUCT_USAGE
+import com.stripe.android.payments.core.injection.PaymentConfigurationModule
 import com.stripe.android.uicore.image.StripeImageLoader
 import com.stripe.android.uicore.isSystemDarkTheme
 import dagger.Binds
@@ -25,7 +24,7 @@ import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import javax.inject.Named
 
-@Module
+@Module(includes = [PaymentConfigurationModule::class])
 internal interface PaymentMethodMessagingModule {
 
     @Binds
@@ -68,17 +67,6 @@ internal interface PaymentMethodMessagingModule {
             application: Application
         ): () -> Boolean {
             return application::isSystemDarkTheme
-        }
-
-        @Provides
-        @Named(PUBLISHABLE_KEY)
-        fun providePublishableKey(
-            configuration: PaymentConfiguration
-        ): () -> String = { configuration.publishableKey }
-
-        @Provides
-        fun paymentConfiguration(application: Application): PaymentConfiguration {
-            return PaymentConfiguration.getInstance(application)
         }
 
         @Provides
