+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/paymentdatacollection/ach/di/USBankAccountFormViewModelModule.kt
@@ -4,36 +4,25 @@ import android.app.Application
 import android.content.Context
 import android.content.res.Resources
 import com.stripe.android.BuildConfig
-import com.stripe.android.PaymentConfiguration
 import com.stripe.android.core.injection.ENABLE_LOGGING
-import com.stripe.android.core.injection.PUBLISHABLE_KEY
 import com.stripe.android.payments.core.injection.PRODUCT_USAGE
+import com.stripe.android.payments.core.injection.PaymentConfigurationModule
 import dagger.Module
 import dagger.Provides
 import javax.inject.Named
 
 @Module(
-    subcomponents = [USBankAccountFormViewModelSubcomponent::class]
+    subcomponents = [USBankAccountFormViewModelSubcomponent::class],
+    includes = [PaymentConfigurationModule::class],
 )
 internal class USBankAccountFormViewModelModule {
     @Provides
     fun providesAppContext(application: Application): Context = application
 
     @Provides
-    fun providePaymentConfiguration(appContext: Context): PaymentConfiguration {
-        return PaymentConfiguration.getInstance(appContext)
-    }
-
-    @Provides
     fun providesResources(appContext: Context): Resources {
         return appContext.resources
     }
-
-    @Provides
-    @Named(PUBLISHABLE_KEY)
-    fun providePublishableKey(
-        appContext: Context
-    ): () -> String = { PaymentConfiguration.getInstance(appContext).publishableKey }
 
     @Provides
     @Named(PRODUCT_USAGE)
