+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/paymentdatacollection/polling/di/PollingViewModelModule.kt
@@ -2,10 +2,9 @@ package com.stripe.android.paymentsheet.paymentdatacollection.polling.di
 
 import android.app.Application
 import android.content.Context
-import com.stripe.android.PaymentConfiguration
 import com.stripe.android.core.injection.ENABLE_LOGGING
-import com.stripe.android.core.injection.PUBLISHABLE_KEY
 import com.stripe.android.payments.core.injection.PRODUCT_USAGE
+import com.stripe.android.payments.core.injection.PaymentConfigurationModule
 import com.stripe.android.paymentsheet.BuildConfig
 import com.stripe.android.paymentsheet.paymentdatacollection.polling.DefaultTimeProvider
 import com.stripe.android.paymentsheet.paymentdatacollection.polling.TimeProvider
@@ -16,7 +15,10 @@ import dagger.Module
 import dagger.Provides
 import javax.inject.Named
 
-@Module(subcomponents = [PollingViewModelSubcomponent::class])
+@Module(
+    subcomponents = [PollingViewModelSubcomponent::class],
+    includes = [PaymentConfigurationModule::class],
+)
 internal interface PollingViewModelModule {
 
     @Binds
@@ -29,17 +31,6 @@ internal interface PollingViewModelModule {
 
         @Provides
         fun providesAppContext(application: Application): Context = application
-
-        @Provides
-        fun providePaymentConfiguration(appContext: Context): PaymentConfiguration {
-            return PaymentConfiguration.getInstance(appContext)
-        }
-
-        @Provides
-        @Named(PUBLISHABLE_KEY)
-        fun providePublishableKey(
-            appContext: Context
-        ): () -> String = { PaymentConfiguration.getInstance(appContext).publishableKey }
 
         @Provides
         @Named(PRODUCT_USAGE)
